+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## vNext
+
+- Add `get_generators` and `get_loads` to NYISO
+
 ## v0.12.0 - Oct 28, 2022
 
 - Can now use `"today"` are value for `end` when querying date range